requirements.txt
alembic==0.8.6
amqp==1.4.9
aniso8601==1.1.0
anyjson==0.3.3
billiard==3.3.0.23
blinker==1.4
celery==3.1.23
docopt==0.6.2
Flask==0.10.1
Flask-Celery-Helper==1.1.0
Flask-Cors==2.1.2
Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.8.0
Flask-OAuthlib==0.9.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
gitdb==0.6.4
GitPython==2.0.5
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
Mako==1.0.4
MarkupSafe==0.23
oauthlib==1.1.1
passlib==1.6.5
Pygments==2.1.3
python-dateutil==2.5.3
python-editor==1.0
pytz==2016.4
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
smmap==0.9.0
SQLAlchemy==1.0.13
tornado==4.3
Werkzeug==0.11.9
WTForms==2.1
